Conflicts:
	apisvc/src/main/java/com/emc/storageos/api/service/impl/resource/BlockConsistencyGroupService.java
	controllersvc/src/main/java/com/emc/storageos/volumecontroller/impl/ControllerUtils.java

+++ vipr-portal/com.iwave.isa.content/src/java/com/emc/sa/service/vipr/file/CreateNfsExportService.java
@@ -4,7 +4,15 @@
  */
 package com.emc.sa.service.vipr.file;
 
-import static com.emc.sa.service.ServiceParams.*;
+import static com.emc.sa.service.ServiceParams.ADVISORY_LIMIT;
+import static com.emc.sa.service.ServiceParams.COMMENT;
+import static com.emc.sa.service.ServiceParams.GRACE_PERIOD;
+import static com.emc.sa.service.ServiceParams.PROJECT;
+import static com.emc.sa.service.ServiceParams.SIZE_IN_GB;
+import static com.emc.sa.service.ServiceParams.SOFT_LIMIT;
+import static com.emc.sa.service.ServiceParams.VIRTUAL_ARRAY;
+import static com.emc.sa.service.ServiceParams.VIRTUAL_POOL;
+import static com.emc.sa.service.ServiceParams.VOLUME_NAME;
 
 import java.net.URI;
 
@@ -29,13 +37,13 @@ public class CreateNfsExportService extends ViPRService {
     protected Double sizeInGb;
     @Param(VOLUME_NAME)
     protected String exportName;
-    @Param(value=SOFT_LIMIT, required=false)
-    protected Integer softLimit;
-    @Param(value=ADVISORY_LIMIT, required=false)
-    protected Integer advisoryLimit;
-    @Param(value=GRACE_PERIOD, required=false)
-    protected Integer gracePeriod;
-    
+    @Param(value = SOFT_LIMIT, required = false)
+    protected Double softLimit;
+    @Param(value = ADVISORY_LIMIT, required = false)
+    protected Double advisoryLimit;
+    @Param(value = GRACE_PERIOD, required = false)
+    protected Double gracePeriod;
+
     @Bindable(itemType = FileStorageUtils.FileExportRule.class)
     protected FileStorageUtils.FileExportRule[] exportRules;
 
@@ -48,11 +56,12 @@ public class CreateNfsExportService extends ViPRService {
 
     @Override
     public void execute() {
-        int tempSoftLimit=(softLimit!=null)?softLimit:0;
-        int tempAdvisoryLimit=(advisoryLimit!=null)?advisoryLimit:0; 
-        int tempGracePeriod=(gracePeriod!=null)?gracePeriod:0;
-        
-        URI fileSystemId = FileStorageUtils.createFileSystem(project, virtualArray, virtualPool, exportName, sizeInGb, tempAdvisoryLimit,tempSoftLimit, tempGracePeriod);
+        int tempSoftLimit = (softLimit != null) ? softLimit.intValue() : 0;
+        int tempAdvisoryLimit = (advisoryLimit != null) ? advisoryLimit.intValue() : 0;
+        int tempGracePeriod = (gracePeriod != null) ? gracePeriod.intValue() : 0;
+
+        URI fileSystemId = FileStorageUtils.createFileSystem(project, virtualArray, virtualPool, exportName, sizeInGb, tempAdvisoryLimit,
+                tempSoftLimit, tempGracePeriod);
         if (exportRules != null) {
             FileStorageUtils.createFileSystemExport(fileSystemId, comment, exportRules[0], null);
             if (exportRules.length > 1) {
